# We can improve slightly UTF-8 validation under some x64 targets.